[msp-smallrom] Update v1.1

+++ sp4.2-software-implementations/FELICS-AE/source/ciphers/.templates/Lilliput_vmsp-smallrom/tbc.c
@@ -30,7 +30,7 @@ static void _compute_round_tweakeys(
     tweakey_state_init(TK, key, tweak);
     tweakey_state_extract(TK, 0, RTK[0]);
 
-    for (uint8_t i=1; i<ROUNDS; i++)
+    for (size_t i=1; i<ROUNDS; i++)
     {
         tweakey_state_update(TK);
         tweakey_state_extract(TK, i, RTK[i]);
@@ -57,7 +57,7 @@ void lilliput_tbc_encrypt(
     RAM_DATA_BYTE RTK[ROUND_TWEAKEY_BYTES];
     tweakey_state_init(TK, key, tweak);
 
-    for (unsigned i=0; i<ROUNDS-1; i++)
+    for (size_t i=0; i<ROUNDS-1; i++)
     {
         tweakey_state_extract(TK, i, RTK);
         nonlinear_and_linear(ciphertext, RTK);
@@ -81,7 +81,7 @@ void lilliput_tbc_decrypt(
     RAM_DATA_BYTE RTK[ROUNDS][ROUND_TWEAKEY_BYTES];
     _compute_round_tweakeys(key, tweak, RTK);
 
-    for (unsigned i=0; i<ROUNDS-1; i++)
+    for (size_t i=0; i<ROUNDS-1; i++)
     {
         nonlinear_and_linear(message, RTK[ROUNDS-1-i]);
         permutation_dec(message);
